+++ pkgs/matchers/binary.go
