+++ pkgs/matchers/document.go
