+++ pkgs/matchers/image.go
